fixed solution
